+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @codebolt/provider
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @codebolt/codeboltjs@5.0.1
+
 ## 2.0.0
 
 ### Patch Changes